null safety mirgations

+++ lib/eu/hydrologis/smash/maps/plugins/fences_plugin.dart
@@ -105,14 +105,14 @@ class CurrentLogPathPainter extends CustomPainter {
       CustomPoint pixelOrigin = map.getPixelOrigin();
 
       CustomPoint mainPosPixel = map.project(center);
-      double mainCenterX = mainPosPixel.x - pixelOrigin.x;
-      double mainCenterY = (mainPosPixel.y - pixelOrigin.y);
+      double mainCenterX = mainPosPixel.x - pixelOrigin.x.toDouble();
+      double mainCenterY = (mainPosPixel.y - pixelOrigin.y.toDouble());
 
       CustomPoint tmpPixelX = map.project(offsetX);
-      double tmpX = tmpPixelX.x - pixelOrigin.x;
+      double tmpX = tmpPixelX.x - pixelOrigin.x.toDouble();
       double rX = (mainCenterX - tmpX).abs();
       CustomPoint tmpPixelY = map.project(offsetY);
-      double tmpY = tmpPixelY.y - pixelOrigin.y;
+      double tmpY = tmpPixelY.y - pixelOrigin.y.toDouble();
       double rY = (mainCenterY - tmpY).abs();
 
       bool hasEnter = fence.onEnter != ENotificationSounds.nosound;
